+++ src/hittable_list.rs
@@ -25,7 +25,7 @@ impl HittableList {
 }
 
 impl Hittable for HittableList {
-  fn hit(&self, r: &Ray, t_min: f32, t_max: f32, rec: &mut HitRecord) -> bool
+  fn hit(&self, r: &Ray, t_min: f64, t_max: f64, rec: &mut HitRecord) -> bool
   where
     Self: Sized,
   {
