+++ src/main.rs
@@ -1,4 +1,4 @@
-use std::f32::INFINITY;
+use std::f64::INFINITY;
 use std::sync::{Arc, Mutex};
 
 extern crate image;
@@ -37,7 +37,7 @@ fn main() {
   // Image
   let aspect_rasio = 16.0 / 9.0;
   let image_width: u32 = 400;
-  let image_height: u32 = (image_width as f32 / aspect_rasio) as u32;
+  let image_height: u32 = (image_width as f64 / aspect_rasio) as u32;
   let samples_per_pixel = 100;
 
   // World
@@ -64,8 +64,8 @@ fn main() {
 
       for _ in 0..pixel_color.samples_per_pixel {
         let mut rng = rand::thread_rng();
-        let u = (i as f32 + rng.gen::<f32>()) / (image_width as f32 - 1.0);
-        let v = (j as f32 + rng.gen::<f32>()) / (image_height as f32 - 1.0);
+        let u = (i as f64 + rng.gen::<f64>()) / (image_width as f64 - 1.0);
+        let v = (j as f64 + rng.gen::<f64>()) / (image_height as f64 - 1.0);
         let r = cam.get_ray(u, v);
         pixel_color.color += ray_color(&r, world.clone());
       }
